fix(i18n): root layout owns html/body — resolves prod hydration cascade
Next.js App Router requires the root app/layout.tsx to render <html> and
<body> directly. Our passthrough (return children) pattern was tolerated
by dev mode but failed production hydration, producing:
  - "Missing required html tags" Next.js warning
  - React error #418 (hydration mismatch)
  - HierarchyRequestError "Only one element on document allowed"
  - NotFoundError removeChild on navigation

Local dev worked; Vercel did not. Root cause was the passthrough root.

Changes:
- app/layout.tsx: owns <html lang> + <body>, loads Poppins, wraps children
  in <Providers>. lang is read via next-intl's getLocale() so /[locale]/*
  gets the resolved locale, and /liff/* + /auth/error fall back to default.
- app/[locale]/layout.tsx: no longer renders <html> — contributes metadata,
  GTM + JSON-LD scripts, noscript iframe, and wraps children in
  NextIntlClientProvider (with locale + messages).
- app/auth/layout.tsx and app/liff/layout.tsx: stripped to passthroughs
  that only contribute scoped metadata/viewport. LIFF retains its
  WebView-tuned viewport (maximumScale: 1, userScalable: false).
- components/layouts/RootShell.tsx deleted — no longer needed now that
  only the root renders <html>. Removes the duplication follow-up entirely.

Verified: npm run typecheck clean, npm run build clean with zero
MISSING_MESSAGE warnings (previously ~200).
